--- a/api/src/services/generate-xform.js
+++ b/api/src/services/generate-xform.js
@@ -5,6 +5,7 @@
 const childProcess = require('child_process');
 const path = require('path');
 const logger = require('../logger');
+const db = require('../db');
 
 const FORM_ROOT_OPEN = '<root xmlns:xf="http://www.w3.org/2002/xforms" xmlns:orx="http://openrosa.org/xforms" xmlns:enk="http://enketo.org/xforms" xmlns:kb="http://kobotoolbox.org/xforms" xmlns:esri="http://esri.com/xforms" xmlns:oc="http://openclinica.org/xforms" xmlns:h="http://www.w3.org/1999/xhtml" xmlns:ev="http://www.w3.org/2001/xml-events" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:jr="http://openrosa.org/javarosa">';
 const MODEL_ROOT_OPEN = '<root xmlns="http://www.w3.org/2002/xforms" xmlns:xf="http://www.w3.org/2002/xforms" xmlns:h="http://www.w3.org/1999/xhtml" xmlns:ev="http://www.w3.org/2001/xml-events" xmlns:xsd="http://www.w3.org/2001/XMLSchema">';
@@ -27,8 +28,8 @@ const transform = (formXml, stylesheet) => {
     xsltproc.stdin.end();
     xsltproc.on('close', code => {
       if (code !== 0 || stderr.length) {
-        console.error('xsltproc stderr output: ');
-        console.error(stderr);
+        logger.error('xsltproc stderr output: ');
+        logger.error(stderr);
         return reject(new Error(`Error transforming xml, xsltproc returned code "${code}"`));
       }
       resolve(stdout);
@@ -70,37 +71,63 @@ const getFormDocs = () => {
     .then(rows => rows.map(row => row.doc));
 };
 
-const getFormAttachment = form => {
-  return getFormDocs()
-    .then(docs => docs.find(doc => doc.internalId === form))
-    .then(doc => {
-      const attachments = (doc && doc._attachments) || {};
-      const name = Object.keys(attachments)
-        .find(name => name === 'xml' || name.endsWith('.xml'));
-      return attachments[name];
-    });
+const getXFormAttachment = doc => {
+  const attachments = (doc && doc._attachments) || {};
+  const name = Object.keys(attachments)
+    .find(name => name === 'xml' || name.endsWith('.xml'));
+  return attachments[name];
 };
 
+const generate = formXml => {
+  return Promise.all([ generateForm(formXml), generateModel(formXml) ])
+    .then(([ form, model ]) => ({ form, model }));
+};
+
+const updateAttachment = (doc, updated, name, type) => {
+  const attachmentData = doc._attachments &&
+                         doc._attachments[name] &&
+                         doc._attachments[name].data &&
+                         doc._attachments[name].data.toString();
+  if (attachmentData === updated) {
+    return false;
+  }
+  doc._attachments[name] = {
+    data: Buffer.from(updated),
+    content_type: type
+  };
+  return true;
+};
+
+const updateAttachmentsIfRequired = (doc, updated) => {
+  const formUpdated = updateAttachment(doc, updated.form, 'form.html', 'text/html');
+  const modelUpdated = updateAttachment(doc, updated.model, 'model.xml', 'text/xml');
+  return formUpdated || modelUpdated;
+}; 
 
 module.exports = {
 
   /**
-   * Generates the xform form and model from the given xml definition.
-   * @param {string} formXml - The xform.
-   * @return {object} An object containing the generated `form` html as
-   *  a string and the generated `model` xml as a string.
+   * Updates the model and form attachments of the given form if necessary. 
+   * @param {string} docId - The db id of the doc defining the form.
    */
-  generate: formXml => {
-    return Promise.all([ generateForm(formXml), generateModel(formXml) ])
-      .then(([ form, model ]) => ({ form, model }));
-  },
-
   update: docId => {
-    return getFormAttachment...
+    return db.medic.get(docId, { attachments: true, binary: true }).then(doc => {
+      const xform = getXFormAttachment(doc);
+      if (!xform) {
+        return Promise.reject({ message: `Form "${docId}" not found`, code: 404 });
+      }
+      return module.exports._generate(xform.data.toString()).then(updated => {
+        if (updateAttachmentsIfRequired(doc, updated)) {
+          return db.medic.put(doc);
+        }
+      });
+    });
   },
 
   updateAll: () => {
     
-  }
+  },
+
+  _generate: generate
 
 };