--- a/api/src/services/generate-xform.js
+++ b/api/src/services/generate-xform.js
@@ -17,6 +17,7 @@ const FORM_STYLESHEET = getStylesheetPath('openrosa2html5form.xsl');
 const MODEL_STYLESHEET = getStylesheetPath('openrosa2xmlmodel.xsl');
 
 const transform = (formXml, stylesheet) => {
+  // TODO test if xsltproc exists on startup and return useful error
   return new Promise((resolve, reject) => {
     const xsltproc = childProcess.spawn('xsltproc', [ stylesheet, '-' ]);
     let stdout = '';
@@ -52,14 +53,14 @@ const generateModel = formXml => {
   });
 };
 
-// TODO pull out parts of forms controller as forms service and reuse here
-//    once this is merged: https://github.com/medic/medic/pull/5656
-const isXMLForm = doc => {
-  return Object.keys(doc && doc._attachments || {})
-         .some(name => name === 'xml' || name.endsWith('.xml'));
+const getEnketoForm = doc => {
+  const collect = doc.context && doc.context.collect;
+  return !collect && getXFormAttachment(doc);
 };
 
-const getFormDocs = () => {
+// TODO pull out parts of forms controller as forms service and reuse here
+//    once this is merged: https://github.com/medic/medic/pull/5656
+const getForms = () => {
   const options = {
     key: ['form'],
     include_docs: true,
@@ -67,8 +68,7 @@ const getFormDocs = () => {
     binary: true,
   };
   return db.medic.query('medic-client/doc_by_type', options)
-    .then(response => response.rows.filter(row => isXMLForm(row.doc)))
-    .then(rows => rows.map(row => row.doc));
+    .then(response => response.rows.map(row => row.doc));
 };
 
 const getXFormAttachment = doc => {
@@ -102,7 +102,37 @@ const updateAttachmentsIfRequired = (doc, updated) => {
   const formUpdated = updateAttachment(doc, updated.form, 'form.html', 'text/html');
   const modelUpdated = updateAttachment(doc, updated.model, 'model.xml', 'text/xml');
   return formUpdated || modelUpdated;
-}; 
+};
+
+const updateAttachments = (accumulator, doc) => {
+  return accumulator.then(results => {
+    const form = getEnketoForm(doc);
+    if (!form) {
+      results.push(null); // not an enketo form - no update required
+      return results;
+    }
+    logger.debug(`Generating html and xml model for enketo form "${doc._id}"`);
+    return module.exports._generate(form.data.toString()).then(result => {
+      results.push(result);
+      return results;
+    });
+  });
+};
+
+// Returns array of docs that need saving.
+const updateAllAttachments = docs => {
+  // TODO use async.series once this is merged: https://github.com/medic/medic/pull/5751
+  // spawn the child processes in series so we don't smash the server
+  return docs.reduce(updateAttachments, Promise.resolve([])).then(results => {
+    return docs.filter((doc, i) => {
+      const result = results[i];
+      if (!result) {
+        return false;
+      }
+      return updateAttachmentsIfRequired(doc, result);
+    });
+  });
+};
 
 module.exports = {
 
@@ -111,21 +141,39 @@ module.exports = {
    * @param {string} docId - The db id of the doc defining the form.
    */
   update: docId => {
-    return db.medic.get(docId, { attachments: true, binary: true }).then(doc => {
-      const xform = getXFormAttachment(doc);
-      if (!xform) {
-        return Promise.reject({ message: `Form "${docId}" not found`, code: 404 });
-      }
-      return module.exports._generate(xform.data.toString()).then(updated => {
-        if (updateAttachmentsIfRequired(doc, updated)) {
+    return db.medic.get(docId, { attachments: true, binary: true })
+      .then(doc => updateAllAttachments([ doc ]))
+      .then(docs => {
+        const doc = docs.length && docs[0];
+        if (doc) {
+          logger.info(`Updating form with ID ${docId}`);
           return db.medic.put(doc);
+        } else {
+          logger.info(`Form with ID ${docId} does not need to be updated.`);
         }
       });
-    });
   },
 
   updateAll: () => {
-    
+    return getForms()
+      .then(docs => {
+        if (!docs.length) {
+          return [];
+        }
+        return updateAllAttachments(docs);
+      })
+      .then(toSave => {
+        logger.info(`Updating ${toSave.length} enketo form${toSave.length === 1 ? '' : 's'}`);
+        if (!toSave.length) {
+          return;
+        }
+        return db.medic.bulkDocs(toSave).then(results => {
+          if (results.some(result => !result.ok)) {
+            throw new Error('Failed to save updated xforms to the database');
+          }
+        });
+      });
+
   },
 
   _generate: generate