… .travis.yml
… api/package.json
… api/src/xsl/openrosa2html5form.xsl
… api/tests/mocha/services/forms.spec.js
… api/tests/mocha/services/generate-xform.spec.js
… api/tests/mocha/services/xforms/multimedia/form.html
… api/tests/mocha/services/xforms/multimedia/model.xml
… api/tests/mocha/services/xforms/multimedia/xform.xml
… api/tests/mocha/services/xforms/simple/form.html
… api/tests/mocha/services/xforms/simple/model.xml
… api/tests/mocha/services/xforms/simple/xform.xml
… grunt/service-worker.js
… package.json
… tests/e2e/service-worker.js
… webapp/package.json
… webapp/src/js/controllers/inbox.js
… webapp/src/js/services/enketo.js
… webapp/src/js/services/index.js
… webapp/src/js/services/xml-forms.js
… webapp/src/js/services/xslt.js
… webapp/tests/karma/unit/services/enketo.js