# Description

Generate the "model.xml" and "form.html" files on the server.

This simplifies and improves performance on the client. The first load will be quicker as
no xslt stage is required. In addition we no longer need to cache the result so memory
usage is reduced.

This also means we can use an (almost) stock enketo-xslt which makes it much easier to
keep on the latest version of enketo.

It also introduces a new runtime dependency on the `xsltproc` command.

medic/medic#4450

# Code review items

- Readable: Concise, well named, follows the [style guide](https://github.com/medic/medic-docs/blob/master/development/style-guide.md), documented if necessary.
- Documented: Configuration and user documentation on [medic-docs](https://github.com/medic/medic-docs/)
- Tested: Unit and/or e2e where appropriate
- Internationalised: All user facing text
- Backwards compatible: Works with existing data and configuration or includes a migration. Any breaking changes documented in Changes.md.

# License

The software is provided under AGPL-3.0. Contributions to this project are accepted under the same license.