@garethbowen 2019-09-04
> I'm interested if you have any performance metrics (though I suppose on the client it would be just removing the first, slower form render)

I timed the Enketo.render function on my main PC (serveo wasn't playing ball today) 3 times and averaged the result:

| | Before | After |
|---|---|---|
| First load | 835ms | 520ms |
| Subsequent load | 451ms | 520ms |

So more or less what you'd expect - executing the xslt is slow, but loading from memory is faster than loading from pouch.

I also estimate this will save around 240kB of permanently cached strings for each form, so a couple of MB for an average project.

> what happens to clients' service-worker caches when they migrate. Does it drop the XSLT stuff or does it just live there until they do a nuke / refresh?

It drops the XSLT when they refresh to update the SW.